[Network] Add two interceptor to implement cache
To properly implement cache and the application use it properly we
should implement two kinds of okhttp interceptor implementation one to
act over the layer between the application and okhttp instance, the another one
is responsible to act over the layer between okhttp instance and the
endpoint api. They will be configured when we provide an okhttp instance
through our dependency injection implementation.

Signed-off-by: Andre Loureiro <loureiro.andrew@gmail.com>